% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Calculate the standard error for a sample}
\usage{
standard_error(n)
}
\arguments{
\item{n}{numeric vector}
}
\value{
number
}
\description{
Calculate the standard error for a sample
}
\examples{
standard_error(rnorm(20))
}
